backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
sqlalchemy>=2.0,<3
alembic>=1.14,<2
psycopg2-binary>=2.9,<3
pydantic-settings>=2.6,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
python-multipart>=0.0.12,<1
email-validator>=2.2,<3
slowapi>=0.1.9,<1
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-anthropic>=0.3,<0.4
langchain-google-genai>=2,<3
tiktoken>=0.7,<1
langchain-chroma>=0.2,<0.3
chromadb>=0.5,<0.7
sentence-transformers>=3,<4
PyMuPDF>=1.24,<2
python-dotenv>=1.0,<2
pytest>=8,<9
httpx>=0.28,<1
